refactor: move key helpers (splitKey/joinKey/slugifyGroup/displayGroupTitle) to store/keys (B-026)
`services/utils.ts` was a junk-drawer used by every layer. First slice
of the demote: move the snippet-key helpers to `store/` where they
belong (keys are a storage concern, not a service helper).

New `src/store/keys.ts` exports `splitKey`, `joinKey`, `slugifyGroup`,
`displayGroupTitle`. Five import sites updated: `store/snippets.ts`,
`ui/components/Modals.ts`, `ui/components/SnippetsTab.ts`,
`ui/components/community/PackageBrowser.ts`, `ui/utils/group-utils.ts`.

Tests move to `src/store/keys.test.ts` (4 describe blocks, 7 it
blocks — kept verbatim from `services/utils.test.ts`).

No behaviour changes. tsc clean, 325/325 tests green.
