+++ src/services/utils.ts
@@ -64,28 +64,5 @@ export function isBadTrigger(key: string): boolean {
     return false;
 }
 
-export function splitKey(key: string): { group: string; name: string } {
-    const i = key.indexOf("/");
-    return i === -1 ? { group: "", name: key } : { group: key.slice(0, i), name: key.slice(i + 1) };
-}
-
-export function joinKey(group: string, name: string): string {
-    return group ? `${group}/${name}` : name;
-}
-
-export function slugifyGroup(label: string): string {
-    return (label || "")
-        .trim()
-        .normalize("NFKD")
-        .replace(/[\u0300-\u036f]/g, "")
-        .replace(/[^a-zA-Z0-9]+/g, "-")
-        .replace(/^-+|-+$/g, "")
-        .toLowerCase();
-}
-
-export function displayGroupTitle(groupKey: string): string {
-    const last = groupKey.includes("/") ? groupKey.split("/", 1)[0] ?? groupKey : groupKey;
-    return last
-        .replace(/[-_]+/g, " ")
-        .replace(/\b\w/g, (c) => c.toUpperCase());
-}
+// Key helpers (`splitKey`, `joinKey`, `slugifyGroup`, `displayGroupTitle`)
+// moved to `src/store/keys.ts` in 1.1.6 (B-026).
